AIhuihua/requirements.txt
fastapi
uvicorn
aiofiles
pika
websocket-client
requests
requests_toolbelt
